backend/requirements.txt
flask
flask-cors
pandas
python-docx
openpyxl
transformers
torch
spacy
language-tool-python
